+++ disorder_metric.c
@@ -6,24 +6,12 @@
 /*   By: lenivorb <lenivorb@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/27 13:25:59 by lenivorb          #+#    #+#             */
-/*   Updated: 2026/05/27 16:00:40 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 15:25:58 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-/*
-function compute_disorder(stack a):
-	mistakes = 0
-	total_pairs = 0
-	for i from 0 to size(a)-1:
-		for j from i+1 to size(a)-1:
-			total_pairs += 1
-			if a[i] > a[j]:
-				mistakes += 1
-	return mistakes / total_pairs
-
---> see : VI.3.2 Disorder metric (mandatory)
-*/
-
+/* Computes the disorder of stack a
+   (i.e. the quota of pairs of numbers which are in the wrong order) */
 float compute_disorder(t_stack_frame *a)
 {
 	int				mistakes;