+++ output.c
@@ -6,7 +6,7 @@
 /*   By: lenivorb <lenivorb@student.42berlin.d      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/28 11:40:28 by lenivorb          #+#    #+#             */
-/*   Updated: 2026/06/04 17:44:51 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 15:42:13 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,17 +18,18 @@
 
 // --- define ---
 
-void print_bench(t_op_track *track)
+/* Prints out the bench statistics */
+void print_bench(t_stats *stats)
 {
-	ft_printf("[bench] disorder:  %d.%d%%\n", (int)(track -> disorder), \
-	(int)(track -> disorder * 100));
-	ft_printf("[bench] strategy:  %s\n", track -> strategy);
-	ft_printf("[bench] total_ops:  %u\n", track -> total_ops);
+	ft_printf("[bench] disorder:  %d.%d%%\n", (int)(stats -> disorder), \
+	(int)(stats -> disorder * 100));
+	ft_printf("[bench] strategy:  %s\n", stats -> strategy);
+	ft_printf("[bench] total_ops:  %u\n", stats -> total_ops);
 	ft_printf("[bench] sa:  %u  sb:  %u  ss:  %u,  pa:  %u,  pb:  %u\n", \
-	(track -> calls)[0], (track -> calls)[1], (track -> calls)[2], \
-	(track -> calls)[3], (track -> calls)[4]);
+	(stats -> calls)[0], (stats -> calls)[1], (stats -> calls)[2], \
+	(stats -> calls)[3], (stats -> calls)[4]);
 	ft_printf("[bench] ra:  %u  rb:  %u  rr:  %u,  rra:  %u,  rrb:  %u  " \
-    "rrr:  %u\n", (track -> calls)[5], (track -> calls)[6], \
-	(track -> calls)[7], (track -> calls)[8], (track -> calls)[9], \
-	(track -> calls)[10]);
+    "rrr:  %u\n", (stats -> calls)[5], (stats -> calls)[6], \
+	(stats -> calls)[7], (stats -> calls)[8], (stats -> calls)[9], \
+	(stats -> calls)[10]);
 }