+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: rcollet <rcollet@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/26 11:55:43 by rcollet           #+#    #+#             */
-/*   Updated: 2026/06/05 14:34:34 by rcollet          ###   ########.fr       */
+/*   Updated: 2026/06/05 15:19:07 by rcollet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,9 +19,9 @@
 
 typedef struct s_interface
 {
-	t_stack			stacks[2]
-	t_op_tracker	*stats;
-	t_func			func[11];	
+	t_stack	stacks[2]
+	t_stats	*stats;
+	t_func	func[11];	
 }	t_interface;
 
 typedef struct s_funcs
@@ -30,13 +30,12 @@ typedef struct s_funcs
 	char	*name;
 }	t_funcs;
 
-typedef struct s_op_tracker
+typedef struct s_stats
 {
 	unsigned int	calls[11];
 	unsigned int	total_ops;
 	float			disorder;
 	char			*strategy;
-	void			(*print_bench)(s_op_tracker*);
-}					t_op_track;
+}					t_stats;
 
 #endif